Move nav components to components/nav
