+++ src/components/nav/NavItem.svelte
